+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Add `convert_textual_links` to `post_bake` recipe
 * Add `suppress_detailed` and `answer_letter_upper` options into `BakeInjectedExerciseQuestion`
 * Add `suppress_detailed` option into `BakeInjectedExerciseSolution`
+* Use `suppress_detailed` and `answer_letter_upper` options in `business-ethics`
 
 
 ## [v2.30.0] - 2026-01-05